+++ internal/documents/model/template/template.go
@@ -1,0 +1,7 @@
+package template
+
+type ITemplate interface {
+	Init(string)
+	Inject(map[string]interface{}) (string, error)
+	Type() string
+}