+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- Building Memoranda for Record
+- Printing an index of unpublished MRs
+- Building WARNOs for printing
 
 ### Changed